# web/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TestResultIn": {
        "additionalProperties": false,
        "description": "FR-34: a test result, and nothing that could be read back as a count.\n\n`area_code` and `syndrome` say which question the p-value answers; without them\ntwo results for one episode would be indistinguishable. Both are aggregate\nlabels the environmental side already publishes (the area is the zone's\npopulation Group), so neither says anything about a person.",
        "properties": {
          "area_code": {
            "title": "Area Code",
            "type": "string"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "effect_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effect Size"
          },
          "episode_id": {
            "title": "Episode Id",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "n_days": {
            "minimum": 1.0,
            "title": "N Days",
            "type": "integer"
          },
          "p_value": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "P Value",
            "type": "number"
          },
          "syndrome": {
            "title": "Syndrome",
            "type": "string"
          }
        },
        "required": [
          "episode_id",
          "area_code",
          "syndrome",
          "method",
          "p_value",
          "effect_size",
          "n_days",
          "computed_at"
        ],
        "title": "TestResultIn",
        "type": "object"
      },
      "UpstreamSearchIn": {
        "additionalProperties": false,
        "description": "FR-35: an unexplained cluster, reduced to where, when and how surprising.",
        "properties": {
          "area_code": {
            "title": "Area Code",
            "type": "string"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "p_value": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "P Value",
            "type": "number"
          },
          "syndrome": {
            "title": "Syndrome",
            "type": "string"
          }
        },
        "required": [
          "area_code",
          "syndrome",
          "day",
          "p_value",
          "method"
        ],
        "title": "UpstreamSearchIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Upstream Core API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/clinical/episodes": {
      "get": {
        "description": "Episodes a clinical test could still say something about.\n\nThe same rule the CDS card uses: PROBABLE or CONFIRMED, and still inside the\nclinical relevance window. A SUSPECTED episode is one unconfirmed report, and\ntesting hospital data against it would spend the health zone's attention on noise.",
        "operationId": "clinical_episodes_clinical_episodes_get",
        "parameters": [
          {
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "sim"
              ],
              "title": "Stream",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clinical Episodes",
        "tags": [
          "clinical"
        ]
      }
    },
    "/clinical/test-result": {
      "post": {
        "description": "Append a ClinicalTestResult. The payload is the body, verbatim, and nothing more.",
        "operationId": "clinical_test_result_clinical_test_result_post",
        "parameters": [
          {
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "sim"
              ],
              "title": "Stream",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestResultIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clinical Test Result",
        "tags": [
          "clinical"
        ]
      }
    },
    "/clinical/upstream-search": {
      "post": {
        "description": "Ask the environmental side to look upstream of an area it has no episode for.\n\nThe area is resolved back to its zone here, on the environmental side, so the\nrequest carries the node the kernel would search from. An area this catchment\ndoes not know is refused rather than logged as a request nobody can act on.",
        "operationId": "upstream_search_clinical_upstream_search_post",
        "parameters": [
          {
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": "live",
              "enum": [
                "live",
                "sim"
              ],
              "title": "Stream",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpstreamSearchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upstream Search",
        "tags": [
          "clinical"
        ]
      }
    },
    "/episodes": {
      "get": {
        "operationId": "list_episodes_episodes_get",
        "parameters": [
          {
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": "live",
              "title": "Stream",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Episodes",
        "tags": [
          "episodes"
        ]
      }
    },
...
  }
}